srv/version-check-srv/requirements.txt
GitPython==3.1.18
python-dotenv==0.17.1
grpcio-tools==1.38.0
